+++ core/src/main/resources/assets/francois/common/common.ts
@@ -1,0 +1,13 @@
+import {Globals} from "../consts";
+
+export class HttpUtils {
+    public static success(code) {
+        return code >= 200 && code < 300;
+    }
+}
+
+
+declare module Foundation {
+    class Accordian {
+    }
+}
