+++ core/src/main/resources/assets/francois/views/jenkins_iframe.html
@@ -1,0 +1,1 @@
+<iframe name="_jenkins" src="{{currentJenkinsUrl}}"></iframe>
