+++ core/src/main/resources/assets/francois/views/job-create.html
@@ -34,7 +34,7 @@
             <div *ng-if="state.success" class="row">
                 <div class="callout secondary">
                     <h5>Create succeeded</h5>
-                    <a [router-link]="['/Jobs', { templateName: templateName }]" class="button">
+                    <a [router-link]="['/ListJobs', { templateName: templateName }]" class="button">
                         View all jobs for <code>{{templateName}}</code>
                     </a>
                 </div>
