+++ core/src/main/resources/assets/francois/views/template-create.html
@@ -16,7 +16,7 @@
             <div *ng-if="state.success" class="row">
                 <div class="callout secondary">
                     <h5>Template created</h5>
-                    <a [router-link]="['/Templates' ]" class="button">
+                    <a [router-link]="['/ListTemplates' ]" class="button">
                         View all templates
                     </a>
                 </div>
