+++ core/src/main/resources/assets/francois/views/top-nav.html
@@ -6,8 +6,8 @@
                     <img src="/assets/images/paradoxlogo.png" class="pdxlogo"/>
                 </a>
             </li>
-            <li class="menu-text"><a [router-link]="['/Templates']">Francois</a></li>
-            <li><a [router-link]="['/Templates']">All Templates</a></li>
+            <li class="menu-text"><a [router-link]="['/ListTemplates']">Francois</a></li>
+            <li><a [router-link]="['/ListTemplates']">All Templates</a></li>
             <li><a [router-link]="['/CreateTemplate']">New Template</a></li>
         </ul>
     </div>
